requirements.txt
numpy
pandas
scikit-learn
pyyaml
matplotlib
seaborn
joblib
tqdm
pyarrow

# Tree-based / Boosted models
xgboost
lightgbm
catboost

# Imbalance handling
imblearn
imbalanced-learn

# Dashboard / API
flask
plotly

